Fix bad signed/unsigned comparisons
Either by casting, or switching to a more appropriate type
for the variable.

+++ common/rdr/FileInStream.cxx
@@ -58,7 +58,7 @@ int FileInStream::pos()
 
 int FileInStream::overrun(int itemSize, int nItems, bool wait)
 {
-  if (itemSize > sizeof(b))
+  if (itemSize > (int)sizeof(b))
     throw Exception("FileInStream overrun: max itemSize exceeded");
 
   if (end - ptr != 0)
